ERT-747: Added base class Ecl3DFile
