+++ devel/python/python/ert/ecl/__init__.py
@@ -88,6 +88,7 @@ from .fortio import FortIO, openFortIO
 from .ecl_kw import EclKW
 from .ecl_3dkw import Ecl3DKW
 from .ecl_file import EclFile , openEclFile
+from .ecl_3d_file import Ecl3DFile 
 from .ecl_init_file import EclInitFile 
 from .ecl_grid import EclGrid
 from .ecl_region import EclRegion
