+++ devel/python/python/ert/ecl/ecl_init_file.py
@@ -14,29 +14,8 @@
 #  See the GNU General Public License at <http://www.gnu.org/licenses/gpl.html> 
 #  for more details. 
 
-from ert.ecl import EclFile, Ecl3DKW
+from ert.ecl import EclFile, Ecl3DKW , Ecl3DFile
 
 
-class EclInitFile(EclFile):
-    def __init__(self , grid , filename , flags = 0):
-        self.grid = grid
-        super(EclInitFile, self).__init__(filename , flags)
-
-
-    def __getitem__(self , index):
-        return_arg = super(EclInitFile, self).__getitem__( index )
-        if isinstance(return_arg,list):
-            kw_list = return_arg
-        else:
-            kw_list = [ return_arg ]
-
-        # Go through all the keywords and try inplace promotion to Ecl3DKW
-        for kw in kw_list:
-            try:
-                Ecl3DKW.castFromKW( kw , self.grid )
-            except ValueError:
-                pass
-
-                
-        return return_arg
-            
+class EclInitFile(Ecl3DFile):
+    pass
